Renamed DistanceMetricDefaults -> Defaults.hpp
