+++ aikido/tests/ompl/OMPLTestHelpers.hpp
@@ -4,7 +4,7 @@
 #include <gtest/gtest.h>
 #include <aikido/constraint/TestableConstraint.hpp>
 #include <aikido/constraint/dart.hpp>
-#include <aikido/distance/DistanceMetricDefaults.hpp>
+#include <aikido/distance/Defaults.hpp>
 #include <aikido/ompl/AIKIDOGeometricStateSpace.hpp>
 #include <aikido/statespace/GeodesicInterpolator.hpp>
 #include <aikido/statespace/StateSpace.hpp>
