+++ aikido/tests/planner/test_SnapPlanner.cpp
@@ -6,7 +6,7 @@
 #include <aikido/statespace/GeodesicInterpolator.hpp>
 #include <aikido/statespace/dart/MetaSkeletonStateSpace.hpp>
 #include <aikido/constraint/CollisionConstraint.hpp>
-#include <aikido/distance/DistanceMetricDefaults.hpp>
+#include <aikido/distance/Defaults.hpp>
 #include <aikido/constraint/TestableConstraint.hpp>
 #include <dart/dart.h>
 #include <tuple>
